Add workaround to use python-pexpect 2.0 and not the distro version.
SILENT: After significant amount of debugging, I could not find a way
to get it to work, so, this should be a reasonable solution. But
something weird is going on with "python-pexpect 2.4", an interesting
detail is that it adds a space just after the 80th character, and if
calling the setwinsize() to adjust it to 32767 (maximum value), that
space is not generated.


SILENT: svn-revision: 387428